+++ python/CHANGELOG.rst
@@ -7,6 +7,11 @@
  - ``ts.subset()`` produces valid tree sequences even if nodes are shuffled
    out of time order (:user:`hyanwong`, :pr:`2479`, :issue:`2473`)
 
+**Features**
+
+ - The ``ts.nodes`` method now takes an ``order`` parameter so that nodes
+   can be visited in time order (:user:`hyanwong`, :pr:`2471`, :issue:`2370`)
+
 **Changes**
 
  - Single statistics computed with ``TreeSequence.general_stat`` are now